requirements.txt
streamlit
langchain
chromadb
google-generativeai
python-dotenv
